refactor: don't use crate root reexports
