+++ src/arch/aarch64/start/hermit_entry.rs
@@ -4,10 +4,10 @@ use aarch64_cpu::asm::barrier::{SY, dsb};
 use hermit_entry::Entry;
 use hermit_entry::boot_info::RawBootInfo;
 
+use crate::arch::kernel::scheduler::TaskStacks;
+use crate::arch::kernel::{CPU_ONLINE, CURRENT_STACK_ADDRESS};
 use crate::config::KERNEL_STACK_SIZE;
 use crate::env;
-use crate::kernel::scheduler::TaskStacks;
-use crate::kernel::{CPU_ONLINE, CURRENT_STACK_ADDRESS};
 
 /// Entrypoint - Initialize Stack pointer and Exception Table
 #[unsafe(no_mangle)]
