+++ src/arch/riscv64/start/smp.rs
@@ -1,9 +1,9 @@
 use core::arch::naked_asm;
 use core::sync::atomic::Ordering;
 
+use crate::arch::kernel::CURRENT_BOOT_ID;
 use crate::arch::riscv64::kernel::CURRENT_STACK_ADDRESS;
 use crate::config::KERNEL_STACK_SIZE;
-use crate::kernel::CURRENT_BOOT_ID;
 
 #[unsafe(naked)]
 pub unsafe extern "C" fn smp_start(hart_id: usize) -> ! {
